Check if file exists before deleting

+++ nodes.py
@@ -525,7 +525,7 @@ def logging_time(original_fn):
 
 #exp_data_dir = os.path.join(current_directory, "exp_data")
 exp_data_dir = os.path.join(folder_paths.output_directory, "exp_data")
-if not os.path.isdir(exp_data_dir):
+if os.path.exists(exp_data_dir) and not os.path.isdir(exp_data_dir):
     print(f"{exp_data_dir} exists but it is not a directory. Removing it.")
     
     os.remove(exp_data_dir)
